dinero_project/requirements.txt
dj-database-url==0.4.2
Django==1.10
django-tastypie==0.14.0
gunicorn==19.7.1
psycopg2==2.7.3.2
python-dateutil==2.6.1
python-mimeparse==1.6.0
six==1.11.0
whitenoise==3.3.1
